+++ entrypoint/rebootinstances/rebootinstances_suite_test.go
@@ -86,7 +86,7 @@ func cleanEIPs() {
 					return
 				}
 			}(ip)
-			wg.Wait()
 		}
+		wg.Wait()
 	}
 }
